feat(trie): refactor codebase to fix dependency
