+++ chainbase/src/main/java/org/tron/core/state/worldstate/WorldStateTrieStore.java
@@ -1,4 +1,4 @@
-package org.tron.core.db.worldstate;
+package org.tron.core.state.worldstate;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
