+++ chainbase/src/main/java/org/tron/core/store/AccountIndexStore.java
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Component;
 import org.tron.core.capsule.AccountCapsule;
 import org.tron.core.capsule.BytesCapsule;
 import org.tron.core.db.TronStoreWithRevoking;
-import org.tron.core.db.accountstate.StateType;
-import org.tron.core.db.accountstate.WorldStateCallBackUtils;
+import org.tron.core.state.worldstate.StateType;
+import org.tron.core.state.worldstate.WorldStateCallBackUtils;
 
 @Component
 public class AccountIndexStore extends TronStoreWithRevoking<BytesCapsule> {
